domain and encryption
