+++ proxy_tests/proxy/proxy.rs
@@ -1,6 +1,5 @@
 // Copyright 2022 TiKV Project Authors. Licensed under Apache-2.0.
 
-use std::ops::RangeBounds;
 pub use std::{
     collections::HashMap,
     io::Write,
